requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email-validator==2.1.0
PyJWT==2.8.0 
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==2.3.7
WTForms==3.2.1
